+++ R/spiro-package.R
@@ -1,15 +1,15 @@
-#' spiro: Work with data from cardiopulmonary exercise testing
+#' spiro: Manage Data from Cardiopulmonary Exercise Testing
 #'
 #' The spiro package enables a fast workflow with raw files from metabolic
-#' carts/spiroergometry systems. It provides simple tools for data import,
-#' processing, summary and visualization.
+#' carts. It provides simple tools for the import, processing, summary and
+#' visualization of cardiopulmonary exercise test data.
 #'
 #' @section Main functions:
 #' \itemize{
-#'   \item Use \code{\link{spiro}} for import and processing of cardiopulmonary
-#'   exercise testing data.
+#'   \item Use \code{\link{spiro}} for the import and processing of
+#'     cardiopulmonary exercise testing data.
 #'   \item Use \code{\link{spiro_summary}} and \code{\link{spiro_max}} for
-#'     a stepwise summary or the maximum parameter values of the
+#'     a step wise summary or the maximum parameter values of the
 #'     imported data.
 #'   \item Use \code{\link{spiro_plot}} to visualize the data.
 #' }