+++ tests/testthat/test-spiro_get.R
@@ -13,28 +13,28 @@ test_that("device guess turns right", {
 })
 
 test_that("import returns a data frame", {
-  expect_snapshot_output(spiro_import_zan(file))
-  expect_snapshot_output(spiro_import_cosmed(cosmed_file))
-  expect_snapshot_output(spiro_import_cortex(cortex_file))
-  expect_snapshot_output(spiro_import_vyntus(vyntus_file))
+  expect_snapshot_output(spiro_get_zan(file))
+  expect_snapshot_output(spiro_get_cosmed(cosmed_file))
+  expect_snapshot_output(spiro_get_cortex(cortex_file))
+  expect_snapshot_output(spiro_get_vyntus(vyntus_file))
 })
 
 test_that("meta data is imported", {
-  expect_s3_class(attr(spiro_import_zan(file), "info"), "data.frame")
-  expect_s3_class(attr(spiro_import_zan(file), "info")$sex, "factor")
-  expect_snapshot_output(attr(spiro_import_zan(file), "info"))
-  expect_snapshot_output(attr(spiro_import_cosmed(cosmed_file), "info"))
-  expect_snapshot_output(attr(spiro_import_cortex(cortex_file), "info"))
-  expect_snapshot_output(attr(spiro_import_vyntus(vyntus_file), "info"))
+  expect_s3_class(attr(spiro_get_zan(file), "info"), "data.frame")
+  expect_s3_class(attr(spiro_get_zan(file), "info")$sex, "factor")
+  expect_snapshot_output(attr(spiro_get_zan(file), "info"))
+  expect_snapshot_output(attr(spiro_get_cosmed(cosmed_file), "info"))
+  expect_snapshot_output(attr(spiro_get_cortex(cortex_file), "info"))
+  expect_snapshot_output(attr(spiro_get_vyntus(vyntus_file), "info"))
 })
 
 test_that("anonymization works", {
-  expect_snapshot_output(attr(spiro_import(file), "info"))
-  expect_snapshot_output(attr(spiro_import(file, anonymize = FALSE), "info"))
+  expect_snapshot_output(attr(spiro_get(file), "info"))
+  expect_snapshot_output(attr(spiro_get(file, anonymize = FALSE), "info"))
   expect_equal(get_anonid("Jesse", "Owens", "12.09.1913"), "e09d4015")
   expect_equal(get_anonid("Jesse", "Owens"), "15a358c3")
   expect_equal(
-    attr(spiro_import(file), "info")$id,
+    attr(spiro_get(file), "info")$id,
     get_anonid("Simon", "Nolte", "04.10.1998")
   )
 })